+++ packages/dialtone-vue2/common/utils/index.js
@@ -378,6 +378,7 @@ export function capitalizeFirstLetter (str, locale = 'en-US') {
  * @param {string} componentName - the component name
  */
 export function warnIfUnmounted (componentRef, componentName) {
+  if (typeof process !== 'undefined' && process.env.NODE_ENV !== 'test') return;
   if (!componentRef || !(componentRef instanceof HTMLElement) || !document?.body) return;
   if (!document.body.contains(componentRef)) {
     console.warn(`The ${componentName} component is not attached to the document body. This may cause issues.`);