+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [9.94.2](https://github.com/dialpad/dialtone/compare/dialtone/v9.94.1...dialtone/v9.94.2) (2025-01-15)
+
+
+### Bug Fixes
+
+* NO-JIRA fix error process is not defined ([#608](https://github.com/dialpad/dialtone/issues/608)) ([06f7c1f](https://github.com/dialpad/dialtone/commit/06f7c1f89eb42fc9e6405f30457635b890f0bc6e))
+
+## [9.94.1](https://github.com/dialpad/dialtone/compare/dialtone/v9.94.0...dialtone/v9.94.1) (2025-01-14)
+
+
+### Bug Fixes
+
+* NO-JIRA set warnIfUnmounted to only run on test env ([#607](https://github.com/dialpad/dialtone/issues/607)) ([ce14a2d](https://github.com/dialpad/dialtone/commit/ce14a2d7238609d7bd77b659a96477fdd349a535))
+
+# [9.94.0](https://github.com/dialpad/dialtone/compare/dialtone/v9.93.0...dialtone/v9.94.0) (2025-01-14)
+
+
+### Features
+
+* **Utilities:** NO-JIRA add intrinsic size classes for width ([#594](https://github.com/dialpad/dialtone/issues/594)) ([6d68b71](https://github.com/dialpad/dialtone/commit/6d68b71a4ead8de3b2717ca396d52aed9bd7fd61))
+
 # [9.93.0](https://github.com/dialpad/dialtone/compare/dialtone/v9.92.0...dialtone/v9.93.0) (2024-12-31)